+++ src/plonemeeting/portal/core/content/institution.py
@@ -4,6 +4,7 @@ import re
 import requests
 from collective.z3cform.datagridfield.datagridfield import DataGridFieldFactory
 from collective.z3cform.datagridfield.row import DictRow
+from imio.helpers.content import get_vocab
 from plone.app.textfield import RichText
 from plone.dexterity.content import Container
 from plone.autoform import directives
@@ -12,7 +13,7 @@ from plone.supermodel import model
 from plonemeeting.portal.core.config import DEFAULT_CATEGORY_IA_DELIB_FIELD, API_HEADERS, \
     CATEGORY_IA_DELIB_FIELDS_MAPPING_EXTRA_INCLUDE
 from zope import schema
-from zope.interface import Interface
+from zope.interface import Interface, invariant, Invalid
 from zope.interface import implementer
 from zope.schema import ValidationError
 
@@ -242,6 +243,23 @@ class IInstitution(model.Schema):
         default="#cccccc",
         constraint=validate_color_parameters,
     )
+
+    @invariant
+    def categories_mappings_invariant(data):
+        mapped_local_category_id = []
+        local_category_id_errors = set()
+        for row in data.categories_mappings:
+            if row['local_category_id'] in mapped_local_category_id:
+                local_category_id_errors.add(row['local_category_id'])
+            else:
+                mapped_local_category_id.append(row['local_category_id'])
+        if local_category_id_errors:
+            local_category_errors = []
+            local_categories = get_vocab(data.__context__, "plonemeeting.portal.vocabularies.local_categories")
+            for cat_id in local_category_id_errors:
+                local_category_errors.append(local_categories.by_value[cat_id].title)
+            raise Invalid("{}: {}".format(_(u'iA.Delib category mapped more than once'),
+                                          ', '.join(local_category_errors)))
 
 
 @implementer(IInstitution)