+++ src/plonemeeting/portal/core/tests/test_browser_institution.py
@@ -1,7 +1,15 @@
 # -*- coding: utf-8 -*-
+from copy import deepcopy
+
+from plonemeeting.portal.core.content.institution import IInstitution
 from plonemeeting.portal.core.tests.portal_test_case import (
     PmPortalDemoFunctionalTestCase,
 )
+from z3c.form import validator
+
+
+class MockForm(object):
+    pass
 
 
 class TestBrowserInstitution(PmPortalDemoFunctionalTestCase):
@@ -45,3 +53,31 @@ class TestBrowserInstitution(PmPortalDemoFunctionalTestCase):
 
         institution_edit_form.render()
         self.assertFalse(hasattr(self.belleville, "delib_categories"))
+
+    def test_categories_mappings_invariant(self):
+        data = {'categories_mappings': deepcopy(self.belleville.categories_mappings)}
+        invariants = validator.InvariantsValidator(None, None, None, IInstitution, None)
+        self.login_as_manager()
+        institution_edit_form = self.belleville.restrictedTraverse("@@edit")
+        # context is overridden while traversing
+        request = self.portal.REQUEST
+        request.set('PUBLISHED', institution_edit_form)
+        self.assertTupleEqual((), invariants.validate(data))
+        data['categories_mappings'].append({'global_category_id': 'administration',
+                                            'local_category_id': 'administration'})
+        validation = invariants.validate(data)
+        self.assertEqual(1, len(validation))
+        self.assertEqual(str(validation[0]), 'iA.Delib category mapped more than once: Administration générale')
+        data['categories_mappings'].append({'global_category_id': 'police',
+                                            'local_category_id': 'police'})
+        validation = invariants.validate(data)
+        self.assertEqual(1, len(validation))
+        self.assertEqual(str(validation[0]), 'iA.Delib category mapped more than once: '
+                                             'Administration générale, Zone de police')
+        # multiple time the same value returns only once in the message
+        data['categories_mappings'].append({'global_category_id': 'administration',
+                                            'local_category_id': 'administration'})
+        validation = invariants.validate(data)
+        self.assertEqual(1, len(validation))
+        self.assertEqual(str(validation[0]), 'iA.Delib category mapped more than once: '
+                                             'Administration générale, Zone de police')